implementation of polynomial proof
Check1: B, inner matrix as linear combination of the data points (beta X^T). Consistent with the B original matrix.
For each component of the output: Lambda and Q. comparison Lambda with the one I get from the inverse of the elementwise square of the matrix. Inconsistent result (check "comparison between lambda")